feat: generate board on server
- with limited metadata for pieces
