+++ back-end/tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "compilerOptions": {
+      "typeRoots": ["node_modules/@types"]
+    }
+  }
